# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de SalonFlow",
        "description": "API REST para gestión de usuarios, servicios y citas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/AppSalon/public/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/servicios": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener todos los servicios activos",
                "operationId": "adbdd4ea56631241a9d2290a3792f5c6",
                "responses": {
                    "200": {
                        "description": "Lista de servicios activos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear un nuevo servicio",
                "operationId": "321e5961f564d93cc9b349e0035591f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "duracion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Tinte Capilar"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 70000
                                    },
                                    "duracion": {
                                        "type": "integer",
                                        "example": 120
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Aplicación de color completo"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Errores de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/servicios/{id}": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener un servicio por ID",
                "operationId": "7d7836fd39ec87f828973f2d67558742",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Actualizar un servicio",
                "operationId": "77087a9c85f4111765087e0a56b3ce01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Corte Dama Premium"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 60000
                                    },
                                    "duracion": {
                                        "type": "integer",
                                        "example": 70
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Corte personalizado con lavado y secado"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servicio actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Eliminar un servicio",
                "operationId": "7619af7c3945d75dbc2ebe2bb47b503e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}